feat: implemetn path logic for board
