+++ management-app/src/constants.tsx
@@ -6,7 +6,7 @@ import { SignUpForm } from './auth/SignUpForm';
 // import AuthPage from './pages/AuthPage';
 import BoardPage from './pages/BoardPage';
 import EditProfile from './pages/EditProfile.jsx';
-import MainPage from './pages/MainPage';
+import MainPage from './pages/MainPage/MainPage';
 import NotFoundPage from './pages/NotFoundPage';
 import WelcomePage from './pages/WelcomePage';
 import { RouteID, TPath } from './types';
@@ -30,7 +30,7 @@ export const ROUTES: TPath[] = [
   },
   {
     id: RouteID.Board,
-    routePath: '/board',
+    routePath: '/board/:id',
     title: 'Board',
     element: <BoardPage />,
     isShownWhenLoggedIn: true,
