+++ management-app/src/pages/MainPage/components/AddBoardBtn/AddBoardBtn.tsx
@@ -1,4 +1,3 @@
-import { MouseEvent } from 'react';
 import { Grid, Typography } from '@mui/material';
 
 type TProps = {
